modification de minimumStockQuantity de double en int

+++ src/main/java/fr/codesbusters/solidstock/dto/product/PostProductDto.java
@@ -21,7 +21,7 @@ public class PostProductDto {
     private String barCode;
     private String buyPrice;
     private String sellPrice;
-    private double minimumStockQuantity;
+    private int minimumStockQuantity;
     private int supplierId;
     private int vatId;
     private int quantityTypeId;
